feat(0): [java]-[wcferry-mvn]-架构调整
