--- clients/java/wcferry-mvn/src/main/java/com/iamteer/BussinessContext.java
@@ -1,17 +1,0 @@
-package com.iamteer;
-
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Component
-@Data
-@Order(100)
-public class BussinessContext {
-
-    private Client client;
-
-}
